scripts: Fix all shellcheck warnings across build scripts
Add .shellcheckrc with source-path=SCRIPTDIR and external-sources=true
so shellcheck can follow sourced env.sh and lib.sh files.  Suppress
SC2034 in env.sh files since all variables are consumed by sourcing
scripts.

Fix quoting issues (SC2086) in command substitutions, docker volume
flags, function arguments, and glob patterns.  Use ${var:?} guards on
rm -rf (SC2115), add || exit after cd calls (SC2164), and separate
declare from assignment (SC2155).  Disable SC2046 on intentional word-
splitting from gpg_docker_flags and SC2043 on single-element arch loops
kept for future multi-arch support.

+++ scripts/debian/container_build-binary.sh
@@ -12,21 +12,21 @@ function prepare_env {
     mkdir -p "$dir_build"
     if [[ -d "$dir_build/$package_name" ]]; then
         log_info "Removing $dir_build/$package_name"
-        rm -rf "$dir_build/$package_name"
+        rm -rf "${dir_build:?}/$package_name"
     fi
 
     log_info "Creating folder $package_name in $dir_build/"
     mkdir -pv "$dir_build/$package_name"
 
     log_info "Copying source packages to $dir_build/"
-    cp -arv "$dir_artifacts/"*${version}.* "$dir_build/"
+    cp -arv "$dir_artifacts/"*"${version}".* "$dir_build/"
 
     if [[ ! -L "$dir_build/$package_source" ]]; then
         log_info "Missing symlink: $dir_build/$package_source, creating"
         ln -sf "$dir_base/$package_source" "$dir_build/$package_source"
     fi
 
-    cd "$dir_build"
+    cd "$dir_build" || exit
 
     log_info "Extracting source package to $dir_build/$package_name-$version_kernel"
     dpkg-source -x "${package_name}_${version}.dsc"
@@ -36,7 +36,8 @@ declare arch=${1:-}
 declare distro=${2:-}
 declare release=${3:-}
 declare build=${4:-${version_build}}
-declare version="$(get_release_version $distro $release $build)"
+declare version
+version="$(get_release_version "$distro" "$release" "$build")"
 
 declare dir_build="/build"
 declare dir_artifacts="$dir_artifacts/$distro/$release"
@@ -48,7 +49,7 @@ prepare_env
 # We need to update our lists to we can install dependencies correctly
 apt-get update
 
-cd "$dir_build/$package_name-$version_kernel"
+cd "$dir_build/$package_name-$version_kernel" || exit
 mk-build-deps -ir -t 'apt-get -y'
 
 log_info "Building binary package for $release"
@@ -57,4 +58,4 @@ $schedtool dpkg-buildpackage --build=binary
 log_info "Copying binary packages to bind mount: $dir_artifacts/"
 mkdir -p "$dir_artifacts"
 
-cp -arv "$dir_build/"*${version}_${arch}* "$dir_artifacts/"
+cp -arv "$dir_build/"*"${version}"_"${arch}"* "$dir_artifacts/"
