+++ scripts/debian/container_dput-to-ppa.sh
@@ -14,7 +14,7 @@ function prepare_env {
     mkdir -p "$dir_build"
     if [[ -d "$dir_build/$package_name" ]]; then
         log_info "Removing $dir_build/$package_name"
-        rm -rf "$dir_build/$package_name"
+        rm -rf "${dir_build:?}/$package_name"
     fi
 
     if [[ ! -L "$dir_build/$package_source" ]]; then
@@ -27,14 +27,15 @@ declare distro=${1:-}
 declare release=${2:-}
 declare build=${3:-}
 
-declare version="$(get_release_version $distro $release $build)"
+declare version
+version="$(get_release_version "$distro" "$release" "$build")"
 declare conf_dput="$dir_base/configs/.dput.cf"
 declare dir_build="/build"
 declare dir_artifacts="$dir_artifacts/$distro/$release"
 
 prepare_env
 
-cd "$dir_build"
+cd "$dir_build" || exit
 cp -av "$dir_artifacts/${package_name}_${version}"* ./
 
 dput --config "$conf_dput" 'liquorix' "${package_name}_${version}_source.changes"
